test zigzag encoding with min/max int64

+++ proto/encode.go
@@ -17,11 +17,11 @@ func EncodeZigZag(v int64) uint64 {
 }
 
 func encodeZigZag64(v int64) uint64 {
-	return uint64((v << 1) ^ (v >> 63))
+	return (uint64(v) << 1) ^ uint64(v>>63)
 }
 
 func encodeZigZag32(v int32) uint32 {
-	return uint32((v << 1) ^ (v >> 31))
+	return (uint32(v) << 1) ^ uint32(v>>31)
 }
 
 type encodeFunc = func([]byte, unsafe.Pointer, flags) (int, error)
